# Back-End/swagger.yaml
swagger: "2.0"
info:
  title: Passport Management API
  version: 1.0.0
  description: API documentation for Passport Management System

host: https://two0060603-passport-application.onrender.com
basePath: /
schemes:
  - http

paths:

  /user/applications:
    get:
      tags:
        - User
      summary: Get all passport applications for a user
      description: Returns the list of passport applications submitted by a specific user.
      parameters:
        - name: user_id
          in: query
          required: true
          description: The ID of the user whose applications should be fetched.
          type: integer
      responses:
        '200':
          description: List of user's passport applications

        '400':
          description: Missing or invalid user_id parameter

        '403':
          description: Only non-admin users can access this API

        '404':
          description: User not found

        '500':
          description: Server or database error

  /register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [username, password, role, email]
            properties:
              username:
                type: string
              password:
                type: string
              role:
                type: string
                enum: [admin, user]
              email:
                type: string
      responses:
        201:
          description: User registered successfully
        400:
          description: Missing fields
        409:
          description: Username already exists

  /login:
    post:
      tags:
        - Authentication
      summary: User login
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [username, password]
            properties:
              username:
                type: string
              password:
                type: string
      responses:
        200:
          description: Login successful
        401:
          description: Invalid credentials

  /slots:
    get:
      tags:
        - Slots
      summary: Check available slots for a given date
      parameters:
        - name: date
          in: query
          required: true
          type: string
      responses:
        200:
          description: Available slots
        400:
          description: Missing date parameter

  /passport/apply:
    post:
      tags:
        - Passport
      summary: Apply for a passport
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [user_id, full_name, dob, nationality, appointment_date]
            properties:
              user_id:
                type: integer
              full_name:
                type: string
              dob:
                type: string
              nationality:
                type: string
              gender:
                type: string
              place_of_birth:
                type: string
              appointment_date:
                type: string
      responses:
        201:
          description: Application submitted
        400:
          description: Missing fields or no slots
        403:
          description: Forbidden (not a user)

  /admin/applications:
    get:
      tags:
        - Admin
      summary: View all passport applications
      parameters:
        - name: user_id
          in: query
          required: true
          type: integer
      responses:
        200:
          description: List of applications
        403:
          description: Unauthorized (only admin)

  /admin/application/{app_id}/status:
    patch:
      tags:
        - Admin
      summary: Update application status
      parameters:
        - name: app_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            required: [user_id, status]
            properties:
              user_id:
                type: integer
              status:
                type: string
                enum: [pending, approved, rejected, completed]
      responses:
        200:
          description: Status updated
        400:
          description: Invalid status
        403:
          description: Unauthorized
        404:
          description: Application not found

  /admin/kpi:
    get:
      tags:
        - Admin
      summary: Get overall KPI statistics for passport applications
      description: Returns total, approved, rejected, and in-progress application counts.
      parameters:
        - name: user_id
          in: query
          type: integer
          required: true
          description: Admin user ID requesting KPI data.
      responses:
        '200':
          description: KPI statistics retrieved successfully

        '403':
          description: Forbidden — Only admin can access KPI

        '500':
          description: Unexpected server or database error